rename lib into src, working on default malady
